refactor: extract and update character overview example
