+++ src/frontend/apps/showcase/src/main.tsx
@@ -6,13 +6,14 @@ import { Separator, ThemeProvider } from "ui-framework"
 import "./index.css"
 import Preview02Example from "./showcases/preview-02/index.tsx"
 import PreviewExample from "./showcases/preview/index.tsx"
-import WipExample, { Example } from "./showcases/wip/index.tsx"
+import { CharacterOverviewExample } from "./showcases/wip/character-overview-example.tsx"
+import WipExample from "./showcases/wip/index.tsx"
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <ThemeProvider>
       {/* <App /> */}
-      <Example />
+      <CharacterOverviewExample />
       <WipExample />
       <Separator className="my-8" />
       <PreviewExample />
